Add ability to import new asic from script file (v2)
This is meant for NPI support only as publicly released designs
can simply add to the static definitions.

This is used by using the --force (-f) option with a filename that has
a @ prefixed on, e.g.

umr -O bits -f @demo/npi/newchip -lr gfx10

Will read the file 'newchip' and import the ip/reg/bit definitions and then
can proceed as normal producing this output:

[WARNING]: Should use --pci when using create_asic_from_script()
        new_chip.gfx10.mmSUPER_SECRET => 0x12345670
                new_chip.gfx10.mmSUPER_SECRET.enable[0:0]

You will need to use --pci to specify a precise device otherwise it won't be
able to talk to it.

Signed-off-by: Tom St Denis <tom.stdenis@amd.com>

(v2): Added fakerizo demo and updated PCI detection logic to be more robust