Improve NPI support (include/name/family)
Add the ability to include files which is useful for larger asics which
can take time to parse.  Now you can split it up per IP block and then include
multiple IP blocks only when necessary.

Also added the ability to assign family IDs (of the UMR enum sort not kernel
values) and change the name of the ASIC so it's not tied to the filename.

Signed-off-by: Tom St Denis <tom.stdenis@amd.com>