Drop trace sorting time even further.
Further optimizations to the VM trace parser that brings the parse
time of ~3.8M lines of trace data down from >3.5seconds to ~1.08 seconds
on my Carrizo (A12-9800).  The goal is to ensure that parsing large
amounts of mappings (for say large games/compute users) is always somewhat
reasonable.

In this case 3.8M lines represents running unigine-heaven on max settings
for about 4 hours with my Polaris10 card.

Signed-off-by: Tom St Denis <tom.stdenis@amd.com>