Update CMakeList files to limit linked libraries
On some systems, the version of LLVM we want to link against
to perform disassembly may have a huge number of libraries
linked against them. Rather than trying to link against a bunch
of libraries that are not needed, this change brings in
only the libraries that UMR needs and cares about.

Submitted-by: Joseph Greathouse <Joseph.Greathouse@amd.com>
Reviewed-by: Tom St Denis <tom.stdenis@amd.com>