UI library for PM4 decoder
This adds a PM4 stream decoder that feeds a developer supplied
user interface with decoded PM4 packets.  As opposed to the
"--ring" command which dumps to stdout these new routines dump
through a externally provided set of callback functions.  To be
used in creating user interfaces around the decoding functionality.

Signed-off-by: Tom St Denis <tom.stdenis@amd.com>