allow discovery without a debugfs 'name' entry
In certain configurations the debugfs file 'name'
might not be populated with a device entry if
no frame buffer is attached.  For instance in headless
or A+I setups.

In these configurations you can use the syntax

umr -i ${instance} -f ${asicname} ...

To issue debugfs commands (registers/waves/etc) to a device
without 'name' contents.

Signed-off-by: Tom St Denis <tom.stdenis@amd.com>